Implement Deposit modal with hooks extracted from Withdraw
These components are very similar.
