+++ sponsor-dapp-v2/src/lib/routes.js
@@ -15,7 +15,7 @@ const routes = [
   { path: "/ViewPositions", component: ViewPositions, exact: false },
   { path: "/ManagePositions/:tokenAddress", component: ManagePositions, exact: false },
   { path: "/Withdraw/:tokenAddress", component: Withdraw, exact: false },
-  { path: "/Deposit", component: Deposit, exact: false },
+  { path: "/Deposit/:tokenAddress", component: Deposit, exact: false },
   { path: "/Borrow", component: Borrow, exact: false },
   { path: "/Repay", component: Repay, exact: false }
 ];
