+++ sponsor-dapp-v2/src/views/Withdraw.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 import { drizzleReactHooks } from "drizzle-react";
 
@@ -8,79 +8,7 @@ import Header from "components/common/Header";
 import IconSvgComponent from "components/common/IconSvgComponent";
 import { withAddedContract } from "lib/contracts";
 import TokenizedDerivative from "contracts/TokenizedDerivative.json";
-
-function useTextInput() {
-  const [amount, setAmount] = useState("");
-  const handleChangeAmount = event => {
-    // Check if regex number matches
-    if (/^(\s*|\d+)$/.test(event.target.value)) {
-      setAmount(event.target.value);
-    }
-  };
-  return { amount, handleChangeAmount };
-}
-
-function useSendTransactionOnLink(cacheSend, amount, history) {
-  const { drizzle } = drizzleReactHooks.useDrizzle();
-  const { toWei } = drizzle.web3.utils;
-
-  const { account } = drizzleReactHooks.useDrizzleState(drizzleState => ({
-    account: drizzleState.accounts[0]
-  }));
-
-  const { send, status } = cacheSend;
-  const [linkedPage, setLinkedPage] = useState();
-  const handleSubmit = event => {
-    event.preventDefault();
-
-    const linkedPage = event.currentTarget.getAttribute("href");
-    setLinkedPage(linkedPage);
-    send(toWei(amount), { from: account });
-  };
-  // If we've successfully withdrawn, reroute to the linkedPage whose `Link` the user clicked on (currently, this can only
-  // ever be the `ManagePositions` linkedPage).
-  useEffect(() => {
-    if (status === "success" && linkedPage) {
-      history.replace(linkedPage);
-    }
-  }, [status, linkedPage, history]);
-  return handleSubmit;
-}
-
-function useCollateralizationInformation(tokenAddress, changeInShortBalance) {
-  const { drizzle, useCacheCall } = drizzleReactHooks.useDrizzle();
-  const { web3 } = drizzle;
-  const { toBN, toWei } = web3.utils;
-  const data = {};
-  data.derivativeStorage = useCacheCall(tokenAddress, "derivativeStorage");
-  data.nav = useCacheCall(tokenAddress, "calcNAV");
-  data.shortMarginBalance = useCacheCall(tokenAddress, "calcShortMarginBalance");
-
-  if (!Object.values(data).every(Boolean)) {
-    return { ready: false };
-  }
-
-  data.collateralizationRequirement = toBN(data.derivativeStorage.fixedParameters.supportedMove)
-    .add(toBN(toWei("1")))
-    .muln(100);
-
-  data.currentCollateralization = "-- %";
-  data.newCollateralizationAmount = "-- %";
-  const navBn = toBN(data.nav);
-  if (!navBn.isZero()) {
-    data.totalHoldings = navBn.add(toBN(data.shortMarginBalance));
-    data.currentCollateralization = data.totalHoldings.muln(100).div(navBn) + "%";
-    if (changeInShortBalance !== "") {
-      data.newCollateralizationAmount =
-        data.totalHoldings
-          .sub(toBN(toWei(changeInShortBalance)))
-          .muln(100)
-          .div(navBn) + "%";
-    }
-  }
-  data.ready = true;
-  return data;
-}
+import { useTextInput, useSendTransactionOnLink, useCollateralizationInformation } from "lib/custom-hooks";
 
 function Withdraw(props) {
   const { tokenAddress } = props.match.params;
@@ -93,7 +21,11 @@ function Withdraw(props) {
   const { send, status } = useCacheSend(tokenAddress, "withdraw");
   const handleWithdrawClick = useSendTransactionOnLink({ send, status }, withdrawAmount, props.history);
 
-  const data = useCollateralizationInformation(tokenAddress, withdrawAmount);
+  const data = useCollateralizationInformation(
+    tokenAddress,
+    // If non-empty, `withdrawAmount` represents a negative `changeInShortBalance`.
+    withdrawAmount !== "" ? "-" + withdrawAmount : withdrawAmount
+  );
   if (!data.ready) {
     return <div>Loading withdraw data</div>;
   }
